requirement.txt
requests
scipy
torch
datasets
torchvision
transformers
scikit-learn
pandas
wandb
huggingface_hub
Pillow
tensorboardX
tqdm
python-dotenv
